Fix 2 DeprecationWarnings: (1) use numpy.nanmean instead of scipy.stats.nanmean. (2) Cast index to int explicitly

+++ pyglow/get_kpap.py
@@ -23,7 +23,7 @@ def get_kpap(dn):
 
     """
     day_index = ( datetime(dn.year, dn.month, dn.day) - epoch).days
-    hour_index = np.floor(dn.hour/3.)
+    hour_index = int(np.floor(dn.hour/3.))
     kp = geophysical_indices[hour_index, day_index]
     ap = geophysical_indices[hour_index+8, day_index]
     f107  = geophysical_indices[16, day_index]
